pyproject.toml
[tool.poetry]
name = "f1-data-collection"
version = "0.1.0"
description = ""
authors = ["eric.zulficar <ezulfica@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
polars = "^1.20.0"
asyncio = "^3.4.3"
aiohttp = "^3.11.11"
boto3 = "^1.36.2"
numpy = "^2.2.2"
pandas = "^2.2.3"
pyarrow = "^19.0.0"
tqdm = "^4.67.1"
google-cloud-bigquery = "^3.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
